requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Mail
psycopg2-binary
Authlib
stripe
Pillow
